refactor: move files to corresponding directories
