+++ src/main/kotlin/blackjack/model/Card.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.model
 
 data class Card(val rank: Rank, val suit: Suit) {
     override fun toString(): String {
