+++ src/main/kotlin/blackjack/model/GamblerInfo.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.model
 
 @JvmInline
 value class GamblerInfo(val name: String) {
