+++ src/main/kotlin/blackjack/model/Suit.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.model
 
 enum class Suit(val symbol: String) {
     SPADE("♠"),
