+++ src/test/kotlin/blackjack/GamblerInfoTest.kt
@@ -1,5 +1,6 @@
 package blackjack
 
+import blackjack.model.GamblerInfo
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
